refactor: frontend folder structure
