+++ apps/ebms/app/documents/page.tsx
@@ -1,5 +1,6 @@
-import DocumentsPipelinePage from '../../features/documents/components/DocumentsPage';
 
-export default function DocumentsPage() {
-  return <DocumentsPipelinePage />;
+import DocumentsPage from '../../features/documents/DocumentsPage';
+
+export default function Documents() {
+  return <DocumentsPage />;
 }
