+++ apps/ebms/features/sidebar/components/SidebarNavItems.tsx
@@ -5,7 +5,7 @@ import {
 } from '@team-4/shadcn';
 import Link from 'next/link';
 import { SidebarItem } from '../../../lib/sidebar/sidebar.types';
-import type { SidebarTone } from '../sidebarTheme';
+import type { SidebarTone } from './sidebarTheme';
 
 type SidebarNavItemProps = {
   item: SidebarItem;
@@ -13,25 +13,23 @@ type SidebarNavItemProps = {
   tone: SidebarTone;
 };
 
-const navItemToneClasses: Record<SidebarTone, { button: string; badge: string }> =
-  {
-    light: {
-      button:
-        'text-[#4b5563] hover:bg-white/75 hover:text-[#111827] data-[active=true]:bg-white data-[active=true]:text-[#111827] data-[active=true]:shadow-[0_1px_2px_rgba(15,23,42,0.08)]',
-      badge: 'bg-[#dff6e5] text-[#197a33]',
-    },
-    dark: {
-      button:
-        'text-[#737378] hover:bg-[#161616] hover:text-[#d1d1d6] data-[active=true]:bg-[#1d1d1d] data-[active=true]:text-[#f5f5f5]',
-      badge: 'bg-[#163b1f] text-[#6fdc84]',
-    },
-  };
+const navItemToneClasses: Record<
+  SidebarTone,
+  { button: string; badge: string }
+> = {
+  light: {
+    button:
+      'text-[#4b5563] hover:bg-white/75 hover:text-[#111827] data-[active=true]:bg-white data-[active=true]:text-[#111827] data-[active=true]:shadow-[0_1px_2px_rgba(15,23,42,0.08)]',
+    badge: 'bg-[#dff6e5] text-[#197a33]',
+  },
+  dark: {
+    button:
+      'text-[#737378] hover:bg-[#161616] hover:text-[#d1d1d6] data-[active=true]:bg-[#1d1d1d] data-[active=true]:text-[#f5f5f5]',
+    badge: 'bg-[#163b1f] text-[#6fdc84]',
+  },
+};
 
-export function SidebarNavItem({
-  item,
-  isActive,
-  tone,
-}: SidebarNavItemProps) {
+export function SidebarNavItem({ item, isActive, tone }: SidebarNavItemProps) {
   const Icon = item.icon;
   const toneClasses = navItemToneClasses[tone];
